requirements.txt
absl-py
jaconv
jsonlines
numpy
pillow
torch
torchvision
